I should probably be better at commit hygene.

+++ src/ConfigManager.h
@@ -41,7 +41,7 @@ public:
         return 0;
     }
 private:
-    long lastSmackTime;
+    int lastSmackTime;
     int smackCount;
 };
 #endif
